Add files via upload
The files currently attached to this section are the edited files from the catkin_ws. The bash shell script and visioning source code are listed in another directory. 

This program reads data from a file that has been interpreted and written to a readable text file in python. All control executions of the JACO Assistive Robot are executed in python.